backend/requirements.txt
fastapi==0.104.1
uvicorn==0.23.2
pydantic==2.4.2
pydantic-settings==2.0.3
python-dotenv==1.0.0
httpx==0.25.1
sqlalchemy==2.0.23
psycopg2-binary==2.9.9
google-auth==2.23.3
google-auth-oauthlib==1.1.0
google-api-python-client==2.107.0
python-jose==3.3.0
passlib==1.7.4
bcrypt==4.0.1
python-multipart==0.0.6
email-validator==2.1.0
openai==1.3.0
pinecone-client[grpc]==2.2.2
tiktoken==0.5.1
